moved color function into separate header file.

+++ src/vec3.h
@@ -7,33 +7,33 @@
 class vec3 {
     public:
         vec3(0, 0, 0) {}
-        vec3(double e0 , double e1, double e2) : e{e0, e1, e2} {}
+        vec3(float e0 , float e1, float e2) : e{e0, e1, e2} {}
 
-        double x() const {return e[0];}
-        double y() const {return e[1];}
-        double z() const {return e[2];}
+        float x() const {return e[0];}
+        float y() const {return e[1];}
+        float z() const {return e[2];}
         
         vec3& operator+() const {return *this;}
         vec3 operator-() const {return vec3(-e[0], -e[1], -e[2]);}
-        double operator[](int i) const {return e[i];}
-        double& operator[](int i) {return e[i];}
+        float operator[](int i) const {return e[i];}
+        float& operator[](int i) {return e[i];}
 
         inline vec3& operator += (const vec3 &v2);
         inline vec3& operator -= (const vec3 &v2);
         inline vec3& operator *= (const vec3 &v2);
         inline vec3& operator /= (const vec3 &v2);
-        inline vec3& operator *= (const double t);
-        inline vec3& operator /= (const double t);
+        inline vec3& operator *= (const float t);
+        inline vec3& operator /= (const float t);
 
-        inline double length() const {
+        inline float length() const {
             return sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
         }
-        inline double squaredLength() const {
+        inline float squaredLength() const {
             return e[0] * e[0] + e[1]  *e[1] + e[2] * e[2];
         }
         inline void makeUnitVector();
 
-        double e[3];
+        float e[3];
 };
 
 using point3 = vec3;
@@ -50,7 +50,7 @@ inline std::ostream& operator<<(std::ostream &os, const vec3 &t){
 }
 
 inline void vec3::makeUnitVector(){
-    double k = 1.0 / sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
+    float k = 1.0 / sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
     e[0] *= k; e[1] *= k; e[2] *= k;
 }
 
@@ -67,20 +67,20 @@ inline vec3 operator/(const vec3 &v1, const vec3 &v2){
     return vec3(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1], v1.e[2] / v2.e[2]);
 }
 
-inline vec3 operator*(double t, const vec3 &v) {
+inline vec3 operator*(float t, const vec3 &v) {
     return vec3(t * v.e[0], t * v.e[1], t* v.e[2]);
 }
-inline vec3 operator/(double t, const vec3 &v) {
+inline vec3 operator/(float t, const vec3 &v) {
     return vec3(v.e[0] / t, v.e[1] / t, v.e[2] / t);
 }
-inline vec3 operator*(const vec3 &v, double t) {
+inline vec3 operator*(const vec3 &v, float t) {
     return vec3(t * v.e[0], t * v.e[1], t* v.e[2]);
 }
-inline vec3 operator/(const vec3 &v, double t) {
+inline vec3 operator/(const vec3 &v, float t) {
     return vec3(v.e[0] / t, v.e[1] / t, v.e[2] / t);
 }
 
-inline double dot(const vec3 &v1, const vec3 &v2){
+inline float dot(const vec3 &v1, const vec3 &v2){
     return v1.e[0] * v2.e[0] + v1.e[0] * v2.e[1] + v1.e[2] * v2.e[2];
 }
 inline vec3 cross(const vec3 &v1, const vec3 &v2) {
@@ -112,14 +112,14 @@ inline vec3& vec3::operator-=(const vec3 &v){
     return * this;
 }
 
-inline vec3& vec3::operator*=(const double t){
+inline vec3& vec3::operator*=(const float t){
     e[0] *= t;
     e[1] *= t;
     e[2] *= t;
     return * this;
 }
-inline vec3& vec3::operator/=(const double t){
-    double k = 1.0 / t;
+inline vec3& vec3::operator/=(const float t){
+    float k = 1.0 / t;
 
     e[0] *= k;
     e[1] *= k;
